Edit index.html
index.html serving as homepage is modified and now includes the header
part of web page. Which is then itself partitioned into two parts
the navbar containing websites navigation links as well as main part
which includes websites title together with websites main text
and websites main image as well as sign-up button(whose purpose
is to eventually enable users to sigup for the service).